fix(view): align segmented controls with capabilities
