+++ OffshoreBudgeting/Views/AddIncomeFormView.swift
@@ -12,6 +12,8 @@ struct AddIncomeFormView: View {
     // MARK: Environment
     @Environment(\.managedObjectContext) var viewContext   // internal so lifecycle extension can access
     @Environment(\.dismiss) private var dismiss
+    @EnvironmentObject private var themeManager: ThemeManager
+    @Environment(\.platformCapabilities) private var capabilities
 
     // MARK: Inputs
     /// If non-nil, loads and edits an existing Income object.
@@ -124,7 +126,10 @@ struct AddIncomeFormView: View {
                 .pickerStyle(.segmented)
                 .labelsHidden()                         // no inline label column
                 .frame(maxWidth: .infinity)             // <- make the control stretch
-                .controlSize(.large)                    // (optional) nicer tap targets
+                .ubSegmentedControlStyle(
+                    capabilities: capabilities,
+                    accentColor: themeManager.selectedTheme.glassPalette.accent
+                )
                 .accessibilityIdentifier("incomeTypeSegmentedControl")
             }
             .frame(maxWidth: .infinity)                 // <- make the row stretch
