+++ OffshoreBudgeting/Views/HomeView.swift
@@ -683,7 +683,10 @@ struct HomeView: View {
                     .pickerStyle(.segmented)
                     .equalWidthSegments()
                     .frame(maxWidth: .infinity)
-                    .modifier(homeSegmentedControlStyling)
+                    .ubSegmentedControlStyle(
+                        capabilities: capabilities,
+                        accentColor: themeManager.selectedTheme.glassPalette.accent
+                    )
                 }
 
                 // Filter bar (sort options)
@@ -698,7 +701,10 @@ struct HomeView: View {
                     .pickerStyle(.segmented)
                     .equalWidthSegments()
                     .frame(maxWidth: .infinity)
-                    .modifier(homeSegmentedControlStyling)
+                    .ubSegmentedControlStyle(
+                        capabilities: capabilities,
+                        accentColor: themeManager.selectedTheme.glassPalette.accent
+                    )
                 }
 
                 // Always-offer Add button when no budget exists so users can
@@ -728,13 +734,6 @@ struct HomeView: View {
         }
         .ub_ignoreSafeArea(edges: .bottom)
         .ub_hideScrollIndicators()
-    }
-
-    private var homeSegmentedControlStyling: HomeSegmentedControlModifier {
-        HomeSegmentedControlModifier(
-            capabilities: capabilities,
-            accentColor: themeManager.selectedTheme.glassPalette.accent
-        )
     }
 
     private var headerSectionSpacing: CGFloat {
@@ -1317,25 +1316,6 @@ private struct HomeHeaderMinWidthModifier: ViewModifier {
 
     private var minimumWidth: CGFloat {
         RootHeaderActionMetrics.minimumGlassWidth(for: capabilities)
-    }
-}
-
-private struct HomeSegmentedControlModifier: ViewModifier {
-    let capabilities: PlatformCapabilities
-    let accentColor: Color
-
-    func body(content: Content) -> some View {
-#if os(macOS)
-        if capabilities.supportsOS26Translucency {
-            content
-        } else {
-            content
-                .controlSize(.large)
-                .tint(accentColor)
-        }
-#else
-        content
-#endif
     }
 }
